+++ library/core/tests/pattern.rs
@@ -1,4 +1,4 @@
-use std::str::pattern::*;
+use std::pattern::*;
 
 // This macro makes it easier to write
 // tests that do a series of iterations
